Create needed interfaces
Creates interfaces for the various model classes to allow for unit
testing. This commit is likely to cause a lot of conflicts, but the
reduced coupling is arguably worth it.

IBoard allows for the complex movement logic to be completely ignored
while testing other parts of the system. IDiceRoller removes the
randomness associated with using DiceRoller by allowing, for example,
a set sequence of rolls to be specified. IPlayer arguably doesn't
require mocking right now, but it does help reduce the impact of future
changes.